+++ src/main/java/net/nightvision/plugin/project/ProjectsCreateScreen.java
@@ -12,6 +12,8 @@ import net.nightvision.plugin.services.ProjectService;
 import org.jetbrains.annotations.NotNull;
 
 import com.intellij.util.ui.JBUI;
+import com.intellij.ui.components.JBLabel;
+
 import javax.swing.*;
 import java.awt.*;
 
@@ -21,7 +23,7 @@ public class ProjectsCreateScreen extends Screen {
     private JButton cancelButton;
     private JButton createButton;
     private JPanel projectsCreatePanel;
-    private JLabel errorMessage;
+    private JBLabel errorMessage;
 
     public JPanel getProjectsCreatePanel() {
         return projectsCreatePanel;
@@ -32,6 +34,7 @@ public class ProjectsCreateScreen extends Screen {
 
         projectsCreatePanel.setBorder(JBUI.Borders.empty(8));
 
+        makeSelectable(errorMessage);
         errorMessage.setVisible(false);
         backButton.addActionListener(e -> {
             mainWindowFactory.openProjectsPage();